cloud_functions/send_tracking_email/requirements.txt
# Function dependencies, for example:
# package>=version
pytz
jinja2
sendgrid
python_http_client
google-cloud-datastore
